Update convert_csv_to_mt.py
Lines 162 and 203, you have 

bars += self._packUniBar(uniBar) 

which is fine, except unibar isn't declared in that scope. If you look closely, unibar is declared in lines 159 and 200. This means that unibar exists only in those "for" blocks (lines 158-161 and lines 199-202). A fix can be as simple as adding a line above the "for" statement (lines 158 and 199) that says 

unibar = (anything, really) 

Hope that helps =)